requirements.txt
# Base source author: YOLOv5 🚀 by Ultralytics, GPL-3.0 license
# Modified by ☘ Molpaxbio Co., Ltd.
# Author: jaesik won <jaesik.won@molpax.com>
# Created: 10/12/2021
# Version: 0.5
# Modified: 16/12/2021

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# Plotting ------------------------------------
pandas>=1.1.4

# Extras --------------------------------------
thop
awscli>=1.22.21
boto3>=1.20.21
botocore>=1.23
pytz>=2021.1